+++ Mise.Core.Common/Events/IInventoryAppEventFactory.cs
@@ -60,7 +60,7 @@ namespace Mise.Core.Common.Events
 		PARLineItemAddedEvent CreatePARLineItemAddedEvent(IEmployee emp, IBaseBeverageLineItem source, int? quantity, IPAR par);
 		PARLineItemAddedEvent CreatePARLineItemAddedEvent(IEmployee emp, string name, string upc, IEnumerable<ItemCategory> category, 
 			int caseSize, LiquidContainer container, int quantity, IPAR par);
-		PARLineItemQuantityUpdatedEvent CreatePARLineItemQuantityUpdatedEvent(IEmployee emp, IPAR par, Guid lineItemID, int quantity);
+		PARLineItemQuantityUpdatedEvent CreatePARLineItemQuantityUpdatedEvent(IEmployee emp, IPAR par, Guid lineItemID, decimal quantity);
 
 		InventorySectionAddedToRestaurantEvent CreateInventorySectionAddedToRestaurantEvent(IEmployee emp, string newSectionName, bool isDefaultSection, bool allowsPartialBottles);
